requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0